A51TweetDisplay/requirements.txt
A51TweetDisplay==0.1
astroid==2.6.6
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
demoji==1.0.0
distlib==0.3.2
filelock==3.0.12
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.2.0
pylint==2.9.6
PySocks==1.7.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
tweepy==3.10.0
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
wrapt==1.12.1
